requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr<2.0,>=1.3
Babel>=1.3
Flask>=0.10,<1.0
oslo.serialization>=1.4.0               # Apache-2.0
python-neutronclient>=2.3.11,<3
docker-py
